Created Game Model

+++ app/src/main/java/com/nullcrew/Domain/Models/FirmAsteroid.java
@@ -8,7 +8,7 @@ import com.nullcrew.UI.Views.GameView;
 public class FirmAsteroid extends Asteroid {
 	private int lives;
 
-	public FirmAsteroid(int x, int y, int width, int height, double speed) {
+	public FirmAsteroid(double x, double y, int width, int height, double speed) {
 		super(x, y, width, height, speed, Color.CYAN, AsteroidType.Firm);
 		lives = 3;
 	}
