+++ app/src/main/java/com/nullcrew/Domain/Models/GameObject.java
@@ -1,13 +1,13 @@
 package com.nullcrew.Domain.Models;
 
 public abstract class GameObject {
-	protected int x;
-	protected int y;
+	protected double x;
+	protected double y;
 	protected int width;
 	protected int height;
 	private int initial_width;
 	private int initial_height;
-	public GameObject(int x, int y, int width, int height) {
+	public GameObject(double x, double y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
@@ -16,19 +16,19 @@ public abstract class GameObject {
 		initial_height=height;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
